+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - JSON Schemas have their properties ordered in a more natural way
 
+### Fixed
+
+- `ObjectTree`: Scalar values are visualized correctly
+
 ## [2.20.0] - 2025-05-13
 
 ### Added